requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
streamlit==1.29.0
python-dotenv==1.0.0
